[IMP] mail: remove usage of isMessagingInitialized and clean main.js
Part of task-2582313
