+++ addons/mail/static/src/components/notification_alert/notification_alert.xml
@@ -3,7 +3,7 @@
 
     <t t-name="mail.NotificationAlert" owl="1">
         <div class="o_NotificationAlert">
-            <t t-if="env.isMessagingInitialized()">
+            <t t-if="messaging">
                 <center t-if="isNotificationBlocked" class="o_notification_alert alert alert-primary">
                     Odoo Push notifications have been blocked. Go to your browser settings to allow them.
                 </center>
