+++ addons/mail/static/src/utils/test_utils.js
@@ -6,14 +6,14 @@ import {
     addMessagingToEnv,
     addTimeControlToEnv,
 } from '@mail/env/test_env';
-import ChatWindowService from '@mail/services/chat_window_service/chat_window_service';
+import { ChatWindowService } from '@mail/services/chat_window_service/chat_window_service';
 import { MessagingService } from '@mail/services/messaging/messaging';
 import { makeDeferred } from '@mail/utils/deferred/deferred';
-import DialogService from '@mail/services/dialog_service/dialog_service';
+import { DialogService } from '@mail/services/dialog_service/dialog_service';
 import { getMessagingComponent } from '@mail/utils/messaging_component';
 import { nextTick } from '@mail/utils/utils';
-import DiscussWidget from '@mail/widgets/discuss/discuss';
-import MessagingMenuWidget from '@mail/widgets/messaging_menu/messaging_menu';
+import { DiscussWidget } from '@mail/widgets/discuss/discuss';
+import { MessagingMenuWidget } from '@mail/widgets/messaging_menu/messaging_menu';
 import { MockModels } from '@mail/../tests/helpers/mock_models';
 
 import AbstractStorageService from 'web.AbstractStorageService';
